合并完成 －－－－－－ 接下来要解决分块上传的时候, 内存增长过快的BUG －－－－－－ 解决完内存增长过快的BUG,再对整个项目的代码进行review

+++ UpYunSDK2.0/UpYunSDK2.0/Utils/UPMutUpload/UMUUploaderManager.h
@@ -71,8 +71,7 @@
 - (NSMutableArray *)uploadWithFile:(NSData *)fileData
                                   policy:(NSString *)policy
                                signature:(NSString *)signature
-                           progressBlock:(void (^)(float percent,
-                                                   long long requestDidSendBytes))progressBlock
+                           progressBlock:(void (^)(CGFloat percent, int64_t requestDidSendBytes))progressBlock
                            completeBlock:(void (^)(NSError * error,
                                                    NSDictionary * result,
                                                    BOOL completed))completeBlock;
